Added files via upload
There are "n" ticket windows in the railway station. i'th window has Ai tickets available. Price of a ticket is equal to the number of tickets remaining in that window at that time. When "m" tickets have been sold, what's the maximum amount of money the railway station can earn?
Ex. n=2, m=4
in 2 window available tickets are : 2 , 5
2nd window sold 4 tickets so 5+4+3+2=14.


There are "n" ticket windows in the railway station. i'th window has Ai tickets available. Price of a ticket is equal to the number of tickets remaining in that window at that time. When "m" tickets have been sold, what's the maximum amount of money the railway station can earn?
Ex. n=2, m=4
in 2 window available tickets are : 2 , 5
2nd window sold 4 tickets so 5+4+3+2=14.